Split the models directory into one file per model class.
