+++ common/data_refinery_common/models/__init__.py
@@ -1,37 +1,55 @@
+from data_refinery_common.models.api_token import APIToken  # noqa
+from data_refinery_common.models.associations.compendium_result_organism_association import (  # noqa
+    CompendiumResultOrganismAssociation,
+)
+from data_refinery_common.models.associations.downloaderjob_originalfile_association import (  # noqa
+    DownloaderJobOriginalFileAssociation,
+)
+from data_refinery_common.models.associations.experiment_organism_association import (  # noqa
+    ExperimentOrganismAssociation,
+)
+from data_refinery_common.models.associations.experiment_result_association import (  # noqa
+    ExperimentResultAssociation,
+)
+from data_refinery_common.models.associations.experiment_sample_association import (  # noqa
+    ExperimentSampleAssociation,
+)
+from data_refinery_common.models.associations.original_file_sample_association import (  # noqa
+    OriginalFileSampleAssociation,
+)
+from data_refinery_common.models.associations.processorjob_dataset_association import (  # noqa
+    ProcessorJobDatasetAssociation,
+)
+from data_refinery_common.models.associations.processorjob_originalfile_association import (  # noqa
+    ProcessorJobOriginalFileAssociation,
+)
+from data_refinery_common.models.associations.sample_computed_file_association import (  # noqa
+    SampleComputedFileAssociation,
+)
+from data_refinery_common.models.associations.sample_result_association import (  # noqa
+    SampleResultAssociation,
+)
 from data_refinery_common.models.command_progress import (  # noqa
     CdfCorrectedAccession,
     SurveyedAccession,
 )
-from data_refinery_common.models.jobs import (  # noqa
-    DownloaderJob,
-    ProcessorJob,
-    SurveyJob,
-    SurveyJobKeyValue,
-)
-from data_refinery_common.models.models import (  # noqa
-    APIToken,
-    CompendiumResult,
-    CompendiumResultOrganismAssociation,
-    ComputationalResult,
+from data_refinery_common.models.compendium_result import CompendiumResult  # noqa
+from data_refinery_common.models.computational_result import ComputationalResult  # noqa
+from data_refinery_common.models.computational_result_annotation import (  # noqa
     ComputationalResultAnnotation,
-    ComputedFile,
-    Dataset,
-    DownloaderJobOriginalFileAssociation,
-    Experiment,
-    ExperimentAnnotation,
-    ExperimentOrganismAssociation,
-    ExperimentResultAssociation,
-    ExperimentSampleAssociation,
-    OrganismIndex,
-    OriginalFile,
-    OriginalFileSampleAssociation,
-    Pipeline,
-    Processor,
-    ProcessorJobDatasetAssociation,
-    ProcessorJobOriginalFileAssociation,
-    Sample,
-    SampleAnnotation,
-    SampleComputedFileAssociation,
-    SampleResultAssociation,
 )
+from data_refinery_common.models.computed_file import ComputedFile  # noqa
+from data_refinery_common.models.dataset import Dataset  # noqa
+from data_refinery_common.models.experiment import Experiment  # noqa
+from data_refinery_common.models.experiment_annotation import ExperimentAnnotation  # noqa
+from data_refinery_common.models.jobs.downloader_job import DownloaderJob  # noqa
+from data_refinery_common.models.jobs.processor_job import ProcessorJob  # noqa
+from data_refinery_common.models.jobs.survey_job import SurveyJob  # noqa
+from data_refinery_common.models.jobs.survey_job_key_value import SurveyJobKeyValue  # noqa
 from data_refinery_common.models.organism import Organism  # noqa
+from data_refinery_common.models.organism_index import OrganismIndex  # noqa
+from data_refinery_common.models.original_file import OriginalFile  # noqa
+from data_refinery_common.models.pipeline import Pipeline  # noqa
+from data_refinery_common.models.processor import Processor  # noqa
+from data_refinery_common.models.sample import Sample  # noqa
+from data_refinery_common.models.sample_annotation import SampleAnnotation  # noqa
