+++ common/data_refinery_common/models/sample_annotation.py
@@ -1,0 +1,37 @@
+from django.contrib.postgres.fields import JSONField
+from django.db import models
+from django.utils import timezone
+
+from data_refinery_common.models.managers import PublicObjectsManager
+
+
+class SampleAnnotation(models.Model):
+    """ Semi-standard information associated with a Sample """
+
+    class Meta:
+        db_table = "sample_annotations"
+        base_manager_name = "public_objects"
+
+    # Managers
+    objects = models.Manager()
+    public_objects = PublicObjectsManager()
+
+    # Relations
+    sample = models.ForeignKey("Sample", blank=False, null=False, on_delete=models.CASCADE)
+
+    # Properties
+    data = JSONField(default=dict)
+    is_ccdl = models.BooleanField(default=False)
+
+    # Common Properties
+    is_public = models.BooleanField(default=True)
+    created_at = models.DateTimeField(editable=False, default=timezone.now)
+    last_modified = models.DateTimeField(default=timezone.now)
+
+    def save(self, *args, **kwargs):
+        """ On save, update timestamps """
+        current_time = timezone.now()
+        if not self.id:
+            self.created_at = current_time
+        self.last_modified = current_time
+        return super(SampleAnnotation, self).save(*args, **kwargs)
